move stripe.js into utils folder
